+++ android/app/src/main/kotlin/com/northabyss/harmonymusic/Equalizer.kt
@@ -1,4 +1,4 @@
-package com.anandnet.harmonymusic
+package com.northabyss.harmonymusic
 
 import android.app.Activity
 import android.content.Context