+++ android/app/src/main/kotlin/com/northabyss/harmonymusic/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.anandnet.harmonymusic
+package com.northabyss.harmonymusic
 
 import com.ryanheise.audioservice.AudioServiceActivity