+++ lib/ui/home.dart
@@ -28,16 +28,18 @@ class Home extends StatelessWidget {
     final homeScreenController = Get.find<HomeScreenController>();
     final size = MediaQuery.of(context).size;
     final isWideScreen = size.width > 800;
-    if (!playerController.initFlagForPlayer &&
-        settingsScreenController.isBottomNavBarEnabled.isFalse) {
-      if (isWideScreen) {
-        playerController.playerPanelMinHeight.value =
-            105 + Get.mediaQuery.padding.bottom;
-      } else {
-        playerController.playerPanelMinHeight.value =
-            75 + Get.mediaQuery.padding.bottom;
+    WidgetsBinding.instance.addPostFrameCallback((_) {
+      if (!playerController.initFlagForPlayer &&
+          settingsScreenController.isBottomNavBarEnabled.isFalse) {
+        if (isWideScreen) {
+          playerController.playerPanelMinHeight.value =
+              105 + Get.mediaQuery.padding.bottom;
+        } else {
+          playerController.playerPanelMinHeight.value =
+              75 + Get.mediaQuery.padding.bottom;
+        }
       }
-    }
+    });
     return PopScope(
       canPop: false,
       onPopInvokedWithResult: (didPop, result) async {