+++ lib/ui/widgets/modification_list.dart
@@ -46,10 +46,7 @@ class ModificationList extends StatelessWidget {
                   ),
                 ),
             itemCount: items.length,
-            onReorder: (old_, new_) {
-              if (old_ < new_) {
-                new_--;
-              }
+            onReorderItem: (old_, new_) {
               final list = items.toList();
               final item = list.removeAt(
                 old_,