+++ test.dart
@@ -1,2 +1,0 @@
-import 'package:flutter_distributor/flutter_distributor.dart';
-void main() {}